GPLR-143 Datenbankfunktionen Refactoring (#38)
Alle Funktionen (bis auf InitializeDatabase) sind jetzt in einer
"tour/stage/locationService.ts" Datei im Ordner services/data

- [x] Wenn nötig, refactored? (Pfadfinderregel)
- [ ] Tests ausreichend vorhanden?
- [x] Code ausreichend kommentiert / dokumentiert?
- [x] Dark-Mode beachtet?
- [ ] PR in Discord gepostet?

---------
