+++ model/model.ts
@@ -17,10 +17,19 @@ class Tour extends Model {
   @field("finished_at") finishedAt?: number;
 
   //@ts-ignore
-  @writer async addStage(title: string, startedAt?: number) {
+  @writer async addStage(
+    title: string,
+    startedAt?: number,
+    finishedAt?: number,
+    active: boolean = false,
+  ) {
     return this.collections.get<Stage>("stages").create((stage) => {
       stage.title = title;
       stage.startedAt = startedAt ?? Date.now();
+      if (finishedAt) {
+        stage.finishedAt = finishedAt;
+      }
+      stage.isActive = active;
       stage.tour.set(this);
     });
   }
